requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
requests
authlib
coverage
pytest
pytest-django
factory-boy
django-filter
celery[redis]
africastalking==2.0
dj-database-url
mozilla-django-oidc
drf-spectacular
django-cors-headers
PyJWT>=2.8.0
djangorestframework-simplejwt>=5.5.1
